+++ src/vuex/index.js
@@ -8,7 +8,7 @@ import kyc from './kyc.module'
 import keyValue from './key-value.module'
 
 import { movementsHistoryModule } from '@/vue/modules/movements-history/store'
-import { issuancesExplorerModule } from '@/vue/modules/issuances-explorer/store'
+import { issuanceExplorerModule } from '@/vue/modules/issuance-explorer/store'
 
 import { vuexTypes } from '@/vuex/types'
 import { sessionStoragePlugin } from './plugins/session-storage'
@@ -42,7 +42,7 @@ const store = new Vuex.Store({
     keyValue,
     // namespaced local modules (used by a specific set of components)
     'movements-history': movementsHistoryModule,
-    'issuances-explorer': issuancesExplorerModule,
+    'issuance-explorer': issuanceExplorerModule,
   },
   plugins: [sessionStoragePlugin],
 })